cut out fidget specific data and trimmed config
