--- src/config/content/nouns.content.ts
@@ -1,215 +1,0 @@
-export const nounsContent = {
-  faq: [
-    {
-      question: "What is a Noun?",
-      answer: "A Noun is a one-of-a-kind 32x32 pixel art character created daily as part of the Nouns project. Each Noun is randomly generated with traits: backgrounds, heads, glasses, body, and accessory. Each Noun is stored on the Ethereum blockchain. Beyond the art, owning a Noun gives you membership in Nouns DAO, a community that manages a shared treasury to fund creative and impactful projects."
-    },
-    {
-      question: "How do I get a Noun?",
-      answer: "You can get a Noun by bidding on the daily auction at nouns.com, buying one from a marketplace, or redeeming $NOUNS tokens (1,000,000 $NOUNS = 1 Noun)."
-    },
-    {
-      question: "What is Nouns DAO?",
-      answer: "Nouns DAO is a decentralized autonomous organization that manages a shared treasury. Noun holders can propose and vote on how to use the treasury to fund projects that benefit the Nouns community and the world."
-    },
-    {
-      question: "How does governance work?",
-      answer: "Noun holders can create proposals for how to use the treasury. Proposals that receive enough votes become executable transactions that can spend from the treasury. Each Noun represents one vote."
-    },
-    {
-      question: "What can I do with my Noun?",
-      answer: "You can use your Noun as your profile picture, participate in governance, create and vote on proposals, join the community, and more. Nouns are also valuable digital collectibles that can be traded."
-    }
-  ],
-  learning: [
-    {
-      title: "Nouns Governance: How Nouns DAO backs proposals",
-      href: "https://www.nouns.com/learn/nouns-governance",
-      image: "/learn/governance.png",
-    },
-    {
-      title: "Noggle the Glossary of Nouns DAO",
-      href: "https://www.nouns.com/learn/noggle-the-glossary",
-      image: "/learn/noggles.png",
-    },
-    {
-      title: "This is Nouns 101",
-      href: "https://www.nouns.com/learn/this-is-nouns-101",
-      image: "/learn/what-are-nouns.png",
-    },
-  ],
-  tutorial: {
-    title: "Welcome to Nounspace!",
-    text: `### 🖌️ Click the paintbrush in the bottom-left corner to open Customization Mode
-
-### Add Fidgets
-1. Click the blue **+** button.
-2. Drag a Fidget to an open spot on the grid.
-3. Click Save
-
-(after saving, scroll down here for more instructions)
-
-### Customize Fidgets
-1. From customization mode, click any Fidget on the grid to open its settings.
-2. Click 'Style' to customize a fidget's look. Any Fidget styles set to "Theme" inherit their look from the Tab's theme.
-
-### Arrange Fidgets
-- **Move:** Drag from the center
-- **Resize:** Drag from an edge or corner
-- **Stash in Fidget Tray:** Click a fidget then click ⇱ to save it for later.
-- **Delete:** Click a fidget then click X it to delete it forever.
-
-### Customize Theme
-- **Templates:** Select a pre-made Theme. Then, customize it further to make it your own.
-- **Style:** Set a background color for the Tab, or set the default styles for all Fidgets on the Tab.
-- **Fonts:** Set the default header and body fonts for Fidgets on the Tab.
-- **Code:** Add HTML/CSS to fully customize the Tab's background, or generate a custom background with a prompt.
-
-### Customize Music
-Add a soundtrack to each Tab. Search for or paste the link to any song or playlist on YouTube, or select a music NFT.
-
-### Homebase vs. Space
-**Your Space** is your public profile that everyone can see.
-**Your Homebase** is a private dashboard that only you can see.
-
-You can use the same tricks and Fidgets to customize them both. Use your **Homebase** to access the content, communities, and functionality you love, and use your **Space** to share the content and functionality you love with your friends.
-
-### Questions or feedback?
-Tag [@nounspacetom](https://nounspace.com/s/nounspacetom) in a cast or join our [Discord](https://discord.gg/H8EYnEmj6q).
-
-### Happy customizing!`,
-    styling: {
-      fontFamily: "Londrina Solid",
-      fontColor: "#073b4c",
-      headingsFontFamily: "Londrina Solid",
-      headingsFontColor: "#2563ea",
-      backgroundColor: "#06d6a0",
-      borderColor: "#ffd166",
-    }
-  },
-  sections: {
-    getANoun: [
-      {
-        key: 'bid',
-        title: "Bid on today's one-of-a-kind Noun and make it yours!",
-        description: "",
-        buttonLabel: "Bid now",
-        href: "https://www.nouns.com/nouns",
-        image: "https://www.nouns.com/feature/join-auction/main.png",
-      },
-      {
-        key: 'shop',
-        title: "Buy a Noun! Shop all major marketplaces in one place.",
-        description: "",
-        buttonLabel: "Shop",
-        href: "https://www.nouns.com/explore?buyNow=1",
-        image: "https://www.nouns.com/feature/shop/main.png",
-      },
-      {
-        key: 'redeem',
-        title: "Collect $NOUNS tokens and redeem them for a Noun.",
-        description: "1,000,000 $NOUNS = 1 Noun",
-        buttonLabel: "Redeem",
-        href: "https://www.nouns.com/convert?tab=redeem",
-        image: "https://www.nouns.com/feature/%24nouns-redeem/main.png",
-      },
-    ],
-    alreadyOwn: [
-      {
-        title: "Instant Swap",
-        description: "Swap your Noun for another in a single trade.",
-        buttonLabel: "Swap",
-        href: "https://www.nouns.com/explore?instantSwap=1",
-        image: "https://www.nouns.com/feature/instant-swap/main.png",
-      },
-      {
-        title: "Treasury Swap",
-        description: "Offer to trade your Noun with one held by the Treasury.",
-        buttonLabel: "Trade",
-        href: "https://www.nouns.com/explore?onlyTreasuryNouns=1",
-        image: "https://www.nouns.com/feature/treasury-swap/main.png",
-      },
-    ],
-    journey: [
-      {
-        title: "Join the Nouns Community",
-        description: "Join the conversation on Farcaster to share ideas and meet other Nouners.",
-        buttonLabel: "Join Nouns on Farcaster",
-        href: "https://warpcast.com/~/channel/nouns",
-        image: "https://www.nouns.com/socials/farcaster.svg",
-        footer: "100k+ Followers",
-      },
-      {
-        title: "Explore Proposals",
-        description: "Discover active proposals and see what the community is funding next.",
-        buttonLabel: "View proposals",
-        href: "https://www.nouns.com/vote",
-        image: "https://www.nouns.com/proposals.svg",
-      },
-    ],
-    fundedProjects: [
-      {
-        title: "Precious Noggles: Recycled Sunglasses",
-        image: "https://www.nouns.com/project/precious-noggles.png",
-        href: "https://www.youtube.com/watch?v=ZGd_mPiTMgQ",
-      },
-      {
-        title: "Hyalinobatrachium Nouns",
-        image: "https://www.nouns.com/project/frog.png",
-        href: "https://explore.nouns.world/hyalinobatrachium-nouns/",
-      },
-      {
-        title: "Quack and Lola",
-        image: "https://www.nouns.com/project/quack.png",
-        href: "https://www.instagram.com/meetquack/",
-      },
-      {
-        title: "Bud Light x Nouns Super Bowl Commercial",
-        image: "https://www.nouns.com/project/bud-light.png",
-        href: "https://explore.nouns.world/bud-light-and-nouns-super-bowl/",
-      },
-      {
-        title: "Dustin Yellin's PERSON, PLACE, or THING",
-        image: "https://www.nouns.com/project/dustin-yellin.png",
-        href: "https://explore.nouns.world/dustin-yellins-person-place-or-thing/",
-      },
-      {
-        title: "John Hamon x Nouns",
-        image: "https://www.nouns.com/project/john-hamon.png",
-        href: "https://explore.nouns.world/john-hamon-and-nouns/",
-      },
-      {
-        title: "The Rise of Blus: A Nouns Movie",
-        image: "https://www.nouns.com/project/rise-of-blus.png",
-        href: "https://nouns.movie/",
-      },
-      {
-        title: "Nouns House: The hub for builders in Sao Paulo",
-        image: "https://www.nouns.com/project/nouns-house.png",
-        href: "https://instagram.com/nouns_house",
-      },
-      {
-        title: "Gnars: Nouns Action Sports Team",
-        image: "https://www.nouns.com/project/gnars.png",
-        href: "https://gnars.com/",
-      },
-      {
-        title: "The Rose Parade: Shark Pickle Cone",
-        image: "https://www.nouns.com/project/rose-parade.png",
-        href: "https://x.com/NounsDoc/status/1836027328616386725",
-      },
-      {
-        title: "Nouns Esports: A New Model for Gaming",
-        image: "https://www.nouns.com/project/e-sports.png",
-        href: "https://nouns.gg/",
-      },
-    ],
-    video: {
-      thumbnail: "https://www.nouns.com/this-is-nouns-video-thumbnail.png",
-      url: "https://www.youtube.com/watch?v=lOzCA7bZG_k",
-    },
-    sampleNounIds: [
-      1n, 10n, 25n, 42n, 99n, 125n, 150n, 208n, 256n, 314n, 365n, 420n, 512n, 777n, 808n, 903n, 1111n, 1337n,
-    ],
-  },
-};
