+++ src/config/fidgets/nouns.fidgets.ts
@@ -20,24 +20,5 @@ export const nounsFidgets = {
     'builderScore',
     'framesV2'
   ],
-  disabled: ['example'],
-  custom: [],
-  defaultLayout: {
-    layoutFidget: "grid",
-    layout: [
-      {
-        w: 12,
-        h: 10,
-        x: 0,
-        y: 0,
-        i: "nounsHome:3a8d7f19-3e77-4c2b-9c7f-1a6f5f5a6f01",
-        minW: 2,
-        maxW: 36,
-        minH: 2,
-        maxH: 36,
-        moved: false,
-        static: false
-      }
-    ]
-  }
+  disabled: ['example']
 };
