+++ src/constants/basedDaos.ts
@@ -1,11 +1,156 @@
-import { loadSystemConfig } from "@/config";
-
-// Load system configuration
-const config = loadSystemConfig();
-
 export const BUILDER_DAO =
   "https://api.goldsky.com/api/public/project_clkk1ucdyf6ak38svcatie9tf/subgraphs/nouns-builder-base-mainnet/stable/gn";
 export const NOUNS_DAO =
   "https://api.goldsky.com/api/public/project_cldf2o9pqagp43svvbk5u3kmo/subgraphs/nouns/prod/gn";
 
-export const DAO_OPTIONS = config.community.daos;
+export const DAO_OPTIONS = [
+  {
+    name: "Nouns DAO",
+    contract: "", // nouns dao does not have a contract address
+    graphUrl: NOUNS_DAO,
+    icon: "/images/nouns_yellow_logo.jpg",
+  },
+  {
+    name: "BASED DAO",
+    contract: "0x10a5676ec8ae3d6b1f36a6f1a1526136ba7938bf",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreiawqtrwmjfylfbwga5ygku3b3dxmlpqy5ozn4ek46vccroveyccom&w=128&q=75",
+  },
+  {
+    name: "Gnars",
+    contract: "0x880fb3cf5c6cc2d7dfc13a993e839a9411200c17",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreif7w4qcje7aikipfgpzjl62fvcff3rtxy43y3s3xvukzur2qgu7cm&w=128&q=75",
+  },
+  {
+    name: "Coppa Nouns",
+    contract: "0x40177254d6f4ced5d09204536f374c381fe325f3",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreibvduf4dwulaq4kkb24rw52rd3vctqdjgqkohz45cgkzdooa4n3tq&w=128&q=75",
+  },
+  {
+    name: "Collective Nouns",
+    contract: "0x220e41499cf4d93a3629a5509410cbf9e6e0b109",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreihte4j5b3trath2xn73psoyjzodalhsj4abvihun7dqcpdez4epoa&w=128&q=75",
+  },
+  {
+    name: "Based Fellas",
+    contract: "0xa5078075b929eabeb02a073edf4243665c9354601",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreid4nwlrqpsxzchtu7sa2nes7slwo5uzsawurhl4yn3t3pzbjqv5je&w=128&q=75",
+  },
+  {
+    name: "Based Associates",
+    contract: "0xeacaead57e40fa0b90e6659f452e2d0bea8b1b33",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeicpjc754sez4tnvngt22nws57wgocn7tfihmuvn4tjkzmj4vpncse&w=128&q=75",
+  },
+  {
+    name: "NounsDAO Africa",
+    contract: "0x614d7503a44e6fd67997f9945bb32d02e8c19431",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreidkedtrbsozscw2lb4hzagee6cia6csjsdqjjwlmx2zabrvhmvvee&w=128&q=75",
+  },
+  {
+    name: "Based Management One",
+    contract: "0xb78b89eb81303a11cc597b4519035079453d8e31",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreihtmu2lgd6aerrizsv3rcke3s2tmorqvt5xqrua3kwlyktt2eqxeu&w=128&q=75",
+  },
+  {
+    name: "MAD BANANA DAO on BASE",
+    contract: "0xf98935b912fec3471ecff0d7cf267b5ebe986a59",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreib4q7kjcthr3viso4ipl3nguvlm4ipr5nmgwffng4arulc7eydvde&w=128&q=75",
+  },
+  {
+    name: "Builder",
+    contract: "0xe8af882f2f5c79580230710ac0e2344070099432",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeieifu437zmpo74odis7pg34ch5svupd5reowpsbb35wih7327mnhy%2Funnamed.png&w=128&q=75",
+  },
+  {
+    name: "OUNCE",
+    contract: "0x8a613cb90ab3b318d4e46d09f260a84b788e206b",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeiaqkwsacyagyo3hny3c5usykerdukh46ti6ftybfqn7fwx7rqlb6m&w=128&q=75",
+  },
+  {
+    name: "Lil Toadz DAO",
+    contract: "0x87175107412c57a2dde6f3f82cab6ec19295efc3",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreie3jkmm4ygb2eumdayx4mh3llmranlj6exbmfajtnfsclwwqpsoni&w=128&q=75",
+  },
+  {
+    name: "member",
+    contract: "0xfbfe187b444798214dd4bbfade369f8dc3864c6a",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreifpmxe7l7gbnp3dfdjqa4ep2whsfszwymltxdiakq6evuejz4bl5a&w=128&q=75",
+  },
+  {
+    name: "Purple",
+    contract: "0x8de71d80ee2c4700bc9d4f8031a2504ca93f7088",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeigofz5ao63vehylvbgx5ikcjfualns4xpx5gmibdojeaydq7khviy%2FFarcaster%2520Purple-8A63D2.png&w=128&q=75",
+  },
+  {
+    name: "PC69",
+    contract: "0xd0dda8e6b354c5f12787aaca936b711c831745ba",
+    graphUrl: BUILDER_DAO,
+  },
+  {
+    name: "the park dao",
+    contract: "0x72b31421a462996f559ffb7fc5dfaca94e754d89",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeibarawmcql724nufftadir52qwj36t7bwycrq7seuajaugxfdriwu%2Fdall__e_2022-09-26_15.05.42_-_golden_gate_park_black_and_white_photo_san_francisco_.png&w=128&q=75",
+  },
+  {
+    name: "Based Mingos",
+    contract: "0x2746eafaac12679b4a0e172daf557e741e58b151",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreifyjhhdtrq5xwrwfjki6dt5xwazz6estacbdu56lriebbq2blktrm&w=128&q=75",
+  },
+  {
+    name: "AcidBinge",
+    contract: "0xf3c3e63994ce37a2888d31ee5675eebb7757b249",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreidihnxtyb7l2wu2prm6vr6awwa5wo47tt6yhhqt26jodxgvdjlwaa&w=128&q=75",
+  },
+  {
+    name: "Najinhas",
+    contract: "0xca511a0b0ca35eda28781e8c5a85124ac91f98dc",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreidtyn6wjtr44hkk2rnpqs7zrbo2dwgu2kuby6z5oupaejllzco7mm&w=128&q=75",
+  },
+  {
+    name: "Nounebels",
+    contract: "0x145d46bf67ab4fadaebf7f8040b43875f95794c4",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreie5jhnfapnfo2fwktees5l6uf5z4rkpo4fvzw5lmd24npsiedbfiu&w=128&q=75",
+  },
+  {
+    name: "SANWeI",
+    contract: "0xc0cf442e851318eab3d710b49cea7d5315b6e967",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeifpmsjqmi2iijvwtevc2owsdohvwouh2asyzccptmry26l4ntprui&w=128&q=75",
+  },
+  {
+    name: "Pioneers",
+    contract: "0xd0d31f743d5f7e8fcf4add1bdb198f07241a4f23",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafybeidjaq56rcqmzh7buy3ddf37bqztuipuca3tsbk52ilnlugbrrmyma&w=128&q=75",
+  },
+  {
+    name: "mferbuilderDAO",
+    contract: "0xc9a73a58e011876aef3f685346afc77fbee58be4",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreicynucppsii7lvkfui4urealvmjrj4eix3jcyw52nbicizny3znz4&w=128&q=75",
+  },
+  {
+    name: "Based Reaper",
+    contract: "0x57092c78e92235acbf2e3e100feb26cd1599f033",
+    graphUrl: BUILDER_DAO,
+    icon: "https://nouns.build/_next/image?url=https%3A%2F%2Fipfs.decentralized-content.com%2Fipfs%2Fbafkreicjzqhdcrpn5zaqxafmsosg3suiyxfmg75sskbdlq5sqn4env5dm4&w=128&q=75",
+  },
+];
