+++ src/constants/nogs.ts
@@ -1,1 +1,6 @@
-export const NOGS_CONTRACT_ADDR = "0xD094D5D45c06c1581f5f429462eE7cCe72215616";
+import { loadSystemConfig } from "@/config";
+
+// Load system configuration
+const config = loadSystemConfig();
+
+export const NOGS_CONTRACT_ADDR = config.community.contracts.nogs;
