+++ src/constants/spaceToken.ts
@@ -1,1 +1,6 @@
-export const SPACE_CONTRACT_ADDR = "0x48c6740bcf807d6c47c864faeea15ed4da3910ab";
+import { loadSystemConfig } from "@/config";
+
+// Load system configuration
+const config = loadSystemConfig();
+
+export const SPACE_CONTRACT_ADDR = config.community.contracts.space;
